The parser's sweeps move out of the 15k-line self-test file
htsselftest.c is twice the size of the next biggest source file, and the tree
already splits a subsystem's self-tests into their own file: htscache_selftest.c
and htsdns_selftest.c, each with a header, listed in Makefile.am, called from
the registry. The three parser sweeps follow that, so htsselftest.c keeps only
the registry rows.

No test changes. The new file needs its row in libhttrack.vcxproj, which
342_msvc-source-lists.test checks.
